+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## WIP
+
+### Fixed
+
+- Fix Icon component on canvas elements (Safari)
 
 ## 3.0.0 beta 1 (January 09, 2017)